fix dependency injection
